+++ libp2p/src/builder/phase/build.rs
@@ -1,11 +1,11 @@
 use std::time::Duration;
 
-#[allow(unused_imports)]
-use super::*;
-
-use crate::SwarmBuilder;
 use libp2p_core::{transport::timeout::TransportTimeout, Transport};
 use libp2p_swarm::Swarm;
+
+#[allow(unused_imports)]
+use super::*;
+use crate::SwarmBuilder;
 
 pub struct BuildPhase<T, B> {
     pub(crate) behaviour: B,