+++ protocols/identify/src/behaviour.rs
@@ -18,28 +18,27 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::handler::{self, Handler, InEvent};
-use crate::protocol::{Info, UpgradeError};
-use libp2p_core::multiaddr::Protocol;
-use libp2p_core::transport::PortUse;
-use libp2p_core::{multiaddr, ConnectedPoint, Endpoint, Multiaddr};
-use libp2p_identity::PeerId;
-use libp2p_identity::PublicKey;
-use libp2p_swarm::behaviour::{ConnectionClosed, ConnectionEstablished, DialFailure, FromSwarm};
-use libp2p_swarm::{
-    ConnectionDenied, DialError, ExternalAddresses, ListenAddresses, NetworkBehaviour,
-    NotifyHandler, PeerAddresses, StreamUpgradeError, THandlerInEvent, ToSwarm,
-    _address_translation,
-};
-use libp2p_swarm::{ConnectionId, THandler, THandlerOutEvent};
-
-use std::collections::hash_map::Entry;
-use std::num::NonZeroUsize;
 use std::{
-    collections::{HashMap, HashSet, VecDeque},
-    task::Context,
-    task::Poll,
+    collections::{hash_map::Entry, HashMap, HashSet, VecDeque},
+    num::NonZeroUsize,
+    task::{Context, Poll},
     time::Duration,
+};
+
+use libp2p_core::{
+    multiaddr, multiaddr::Protocol, transport::PortUse, ConnectedPoint, Endpoint, Multiaddr,
+};
+use libp2p_identity::{PeerId, PublicKey};
+use libp2p_swarm::{
+    behaviour::{ConnectionClosed, ConnectionEstablished, DialFailure, FromSwarm},
+    ConnectionDenied, ConnectionId, DialError, ExternalAddresses, ListenAddresses,
+    NetworkBehaviour, NotifyHandler, PeerAddresses, StreamUpgradeError, THandler, THandlerInEvent,
+    THandlerOutEvent, ToSwarm, _address_translation,
+};
+
+use crate::{
+    handler::{self, Handler, InEvent},
+    protocol::{Info, UpgradeError},
 };
 
 /// Whether an [`Multiaddr`] is a valid for the QUIC transport.
@@ -117,20 +116,20 @@ pub struct Behaviour {
 pub struct Config {
     /// Application-specific version of the protocol family used by the peer,
     /// e.g. `ipfs/1.0.0` or `polkadot/1.0.0`.
-    pub protocol_version: String,
+    protocol_version: String,
     /// The public key of the local node. To report on the wire.
-    pub local_public_key: PublicKey,
+    local_public_key: PublicKey,
     /// Name and version of the local peer implementation, similar to the
     /// `User-Agent` header in the HTTP protocol.
     ///
     /// Defaults to `rust-libp2p/<libp2p-identify-version>`.
-    pub agent_version: String,
+    agent_version: String,
     /// The interval at which identification requests are sent to
     /// the remote on established connections after the first request,
     /// i.e. the delay between identification requests.
     ///
     /// Defaults to 5 minutes.
-    pub interval: Duration,
+    interval: Duration,
 
     /// Whether new or expired listen addresses of the local node should
     /// trigger an active push of an identify message to all connected peers.
@@ -140,19 +139,19 @@ pub struct Config {
     /// i.e. before the next periodic identify request with each peer.
     ///
     /// Disabled by default.
-    pub push_listen_addr_updates: bool,
+    push_listen_addr_updates: bool,
 
     /// How many entries of discovered peers to keep before we discard
     /// the least-recently used one.
     ///
     /// Disabled by default.
-    pub cache_size: usize,
+    cache_size: usize,
 
     /// Whether to include our listen addresses in our responses. If enabled,
     /// we will effectively only share our external addresses.
     ///
     /// Disabled by default.
-    pub hide_listen_addrs: bool,
+    hide_listen_addrs: bool,
 }
 
 impl Config {
@@ -201,6 +200,41 @@ impl Config {
     pub fn with_hide_listen_addrs(mut self, b: bool) -> Self {
         self.hide_listen_addrs = b;
         self
+    }
+
+    /// Get the protocol version of the Config.
+    pub fn protocol_version(&self) -> &str {
+        &self.protocol_version
+    }
+
+    /// Get the local public key of the Config.
+    pub fn local_public_key(&self) -> &PublicKey {
+        &self.local_public_key
+    }
+
+    /// Get the agent version of the Config.
+    pub fn agent_version(&self) -> &str {
+        &self.agent_version
+    }
+
+    /// Get the interval of the Config.
+    pub fn interval(&self) -> Duration {
+        self.interval
+    }
+
+    /// Get the push listen address updates boolean value of the Config.
+    pub fn push_listen_addr_updates(&self) -> bool {
+        self.push_listen_addr_updates
+    }
+
+    /// Get the cache size of the Config.
+    pub fn cache_size(&self) -> usize {
+        self.cache_size
+    }
+
+    /// Get the hide listen address boolean value of the Config.
+    pub fn hide_listen_addrs(&self) -> bool {
+        self.hide_listen_addrs
     }
 }
 
@@ -288,7 +322,8 @@ impl Behaviour {
             .contains(&connection_id)
         {
             // Apply address translation to the candidate address.
-            // For TCP without port-reuse, the observed address contains an ephemeral port which needs to be replaced by the port of a listen address.
+            // For TCP without port-reuse, the observed address contains an ephemeral port which
+            // needs to be replaced by the port of a listen address.
             let translated_addresses = {
                 let mut addrs: Vec<_> = self
                     .listen_addresses
@@ -325,7 +360,7 @@ impl Behaviour {
         }
 
         // outgoing connection dialed with port reuse
-        // incomming connection
+        // incoming connection
         self.events
             .push_back(ToSwarm::NewExternalAddrCandidate(observed.clone()));
     }
@@ -363,7 +398,8 @@ impl NetworkBehaviour for Behaviour {
     ) -> Result<THandler<Self>, ConnectionDenied> {
         // Contrary to inbound events, outbound events are full-p2p qualified
         // so we remove /p2p/ in order to be homogeneous
-        // this will avoid Autonatv2 to probe twice the same address (fully-p2p-qualified + not fully-p2p-qualified)
+        // this will avoid Autonatv2 to probe twice the same address (fully-p2p-qualified + not
+        // fully-p2p-qualified)
         let mut addr = addr.clone();
         if matches!(addr.iter().last(), Some(multiaddr::Protocol::P2p(_))) {
             addr.pop();
@@ -380,7 +416,9 @@ impl NetworkBehaviour for Behaviour {
             self.config.local_public_key.clone(),
             self.config.protocol_version.clone(),
             self.config.agent_version.clone(),
-            addr.clone(), // TODO: This is weird? That is the public address we dialed, shouldn't need to tell the other party?
+            // TODO: This is weird? That is the public address we dialed,
+            // shouldn't need to tell the other party?
+            addr.clone(),
             self.all_addresses(),
         ))
     }