+++ protocols/kad/src/query/peers/closest.rs
@@ -18,13 +18,19 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use super::*;
+use std::{
+    collections::btree_map::{BTreeMap, Entry},
+    num::NonZeroUsize,
+    time::Duration,
+};
 
-use crate::kbucket::{Distance, Key, KeyBytes};
-use crate::{ALPHA_VALUE, K_VALUE};
-use std::collections::btree_map::{BTreeMap, Entry};
-use std::{num::NonZeroUsize, time::Duration};
 use web_time::Instant;
+
+use super::*;
+use crate::{
+    kbucket::{Distance, Key, KeyBytes},
+    ALPHA_VALUE, K_VALUE,
+};
 
 pub(crate) mod disjoint;
 /// A peer iterator for a dynamically changing list of peers, sorted by increasing
@@ -494,12 +500,14 @@ enum PeerState {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::SHA_256_MH;
+    use std::iter;
+
     use libp2p_core::multihash::Multihash;
     use quickcheck::*;
     use rand::{rngs::StdRng, Rng, SeedableRng};
-    use std::iter;
+
+    use super::*;
+    use crate::SHA_256_MH;
 
     fn random_peers<R: Rng>(n: usize, g: &mut R) -> Vec<PeerId> {
         (0..n)