+++ swarm/src/behaviour/peer_addresses.rs
@@ -1,12 +1,10 @@
-use crate::behaviour::FromSwarm;
-use crate::{DialError, DialFailure, NewExternalAddrOfPeer};
+use std::num::NonZeroUsize;
 
 use libp2p_core::Multiaddr;
 use libp2p_identity::PeerId;
-
 use lru::LruCache;
 
-use std::num::NonZeroUsize;
+use crate::{behaviour::FromSwarm, DialError, DialFailure, NewExternalAddrOfPeer};
 
 /// Struct for tracking peers' external addresses of the [`Swarm`](crate::Swarm).
 #[derive(Debug)]
@@ -46,7 +44,6 @@ impl PeerAddresses {
     /// Appends address to the existing set if peer addresses already exist.
     /// Creates a new cache entry for peer_id if no addresses are present.
     /// Returns true if the newly added address was not previously in the cache.
-    ///
     pub fn add(&mut self, peer: PeerId, address: Multiaddr) -> bool {
         match prepare_addr(&peer, &address) {
             Ok(address) => {
@@ -98,16 +95,16 @@ impl Default for PeerAddresses {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use std::io;
 
-    use crate::ConnectionId;
     use libp2p_core::{
         multiaddr::Protocol,
         transport::{memory::MemoryTransportError, TransportError},
     };
-
     use once_cell::sync::Lazy;
+
+    use super::*;
+    use crate::ConnectionId;
 
     #[test]
     fn new_peer_addr_returns_correct_changed_value() {