+++ swarm-test/CHANGELOG.md
@@ -1,7 +1,9 @@
-## 0.4.1
+## 0.5.0
 
 - Add `tokio` runtime support and make `tokio` and `async-std` runtimes optional behind features.
   See [PR 5551].
+  - Update default for idle-connection-timeout to 10s on `SwarmExt::new_ephemeral` methods.
+  See [PR 4967](https://github.com/libp2p/rust-libp2p/pull/4967).
 
 [PR 5551]: https://github.com/libp2p/rust-libp2p/pull/5551